src/ingestion-service/requirements.txt
# Ingestion Service Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
azure-eventhub>=5.11.0
azure-identity>=1.15.0
aiohttp>=3.9.0
httpx>=0.26.0
pydantic>=2.5.0
python-dotenv>=1.0.0
